fix: apply prethrower to all public class methods
